requirements.txt
aiodns==3.0.0
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
backoff==1.11.1
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.6
colorama==0.4.4
colorlog==6.4.1
dnspython==2.1.0
et-xmlfile==1.1.0
idna==3.2
multidict==5.1.0
openpyxl==3.0.9
pycares==4.0.0
pycparser==2.20
PySocks==1.7.1
requests==2.26.0
termcolor==1.1.0
tqdm==4.62.3
typing-extensions==3.10.0.2
urllib3==1.26.7
wincertstore==0.2
yarl==1.6.3
